% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorrxiv_summary.R
\name{biorrxiv_summary}
\alias{biorrxiv_summary}
\title{Retrieve summary statistics for deposits of bioRxiv content}
\usage{
biorrxiv_summary(interval = "m", format = "list", ...)
}
\arguments{
\item{interval}{(character) Return either monthly `m` or yearly `y` data.
Default: `m`}

\item{format}{(character) Return data in list `list`, json `json` or data
frame `df` format. Default: `list`}
}
\description{
Retrieve summary statistics for deposits of bioRxiv content
}
\examples{
# Return a list of monthly deposit statistics
biorrxiv_summary(interval = "m")

# Return data in a data frame
biorrxiv_summary(interval = "m", format = "df")

# Return annual deposit statistics
biorrxiv_summary(interval = "y")

}
