% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorxiv_summary.R
\name{biorxiv_summary}
\alias{biorxiv_summary}
\title{Retrieve summary statistics for deposits of bioRxiv content}
\usage{
biorxiv_summary(interval = "m", format = "list")
}
\arguments{
\item{interval}{(character) Return either monthly \code{m} or yearly \code{y} data.
Default: \code{m}}

\item{format}{(character) Return data in list \code{list}, json \code{json} or data
frame \code{df} format. Default: \code{list}}
}
\description{
Retrieve summary statistics for deposits of bioRxiv content
}
\examples{
\donttest{

# Return a list of monthly deposit statistics
biorxiv_summary(interval = "m")

# Return data in a data frame
biorxiv_summary(interval = "m", format = "df")

# Return annual deposit statistics
biorxiv_summary(interval = "y")
}
}
