% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorxiv_usage.R
\name{biorxiv_usage}
\alias{biorxiv_usage}
\title{Retrieve summary statistics for usage of bioRxiv content}
\usage{
biorxiv_usage(interval = "m", format = "list")
}
\arguments{
\item{interval}{(character) Return either monthly \code{m} or yearly \code{y} data.
Default: \code{m}}

\item{format}{(character) Return data in list \code{list}, json \code{json} or data
frame \code{df} format. Default: \code{list}}
}
\description{
Retrieve summary statistics for usage of bioRxiv content
}
\examples{
\donttest{

# Return a list of monthly usage statistics
biorxiv_usage(interval = "m")

# Return data in a data frame
biorxiv_usage(interval = "m", format = "df")

# Return annual usage statistics
biorxiv_usage(interval = "y")
}
}
