% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorrxiv_summary.R
\name{biorrxiv_summary}
\alias{biorrxiv_summary}
\title{Retrieve summary statistics for deposits of bioRxiv content}
\usage{
biorrxiv_summary(interval = "m", format = "list", ...)
}
\arguments{
\item{interval}{(character) Return either monthly \code{m} or yearly \code{y} data.
Default: \code{m}}

\item{format}{(character) Return data in list \code{list}, json \code{json} or data
frame \code{df} format. Default: \code{list}}
}
\description{
Retrieve summary statistics for deposits of bioRxiv content
}
\examples{
\dontrun{

# Return a list of monthly deposit statistics
biorrxiv_summary(interval = "m")

# Return data in a data frame
biorrxiv_summary(interval = "m", format = "df")

# Return annual deposit statistics
biorrxiv_summary(interval = "y")
}
}
