% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorrxiv_usage.R
\name{biorrxiv_usage}
\alias{biorrxiv_usage}
\title{Retrieve summary statistics for usage of bioRxiv content}
\usage{
biorrxiv_usage(interval = "m", format = "list", ...)
}
\arguments{
\item{interval}{(character) Return either monthly `m` or yearly `y` data.
Default: `m`}

\item{format}{(character) Return data in list `list`, json `json` or data
frame `df` format. Default: `list`}
}
\description{
Retrieve summary statistics for usage of bioRxiv content
}
\examples{
# Return a list of monthly usage statistics
biorrxiv_usage(interval = "m")

# Return data in a data frame
biorrxiv_usage(interval = "m", format = "df")

# Return annual usage statistics
biorrxiv_usage(interval = "y")

}
